Add files via upload
This is the first Java class for creating the User account for those on the shopping platform. We use a String variable to hold customerID for later retrieval in future searching methods and a HashMap to hold user data such as their username and password. This will be vital in finding an account based on their username and password for the UserManagement Class later on. We use simple getters and setters for the different attributes and we store the username and password as their hashcodes to make the customerID. This was a unique idea since each customerID should be different. 

We used simple hashing techniques for the password in the User class with the hashCode() method. This is not meant for secure password hashing, it is simply to store and retrieve the password from the HashMap when we are authenticating or registering user accounts.