Main Class to test User and UserManagement
I am currently testing the methods in User and UserManagement. Right now I have to handle a problem with authenticating the password because right now the authenticating is returning false for both instances of a correct and incorrect log in. I should make a toString method in the User class to see if there is a problem with storing the password first.