ProductCatalog First Version
This Object class will have an ArrayList of any added Product objects from the main class. I decided to go with an ArrayList since it had constant time complexity for adding and removing Product from the ArrayList. For traversing it will have a time complexity of O(n) where n is the number of Product objects in the ArrayList to look through. I currently don't have any search methods but I will implement them after I test the functionality of Product and ProductCatalog in the Main java class.