Testing Search Methods
Each search method uses a for loop to loop through the ArrayList in each product and comparing its attributes to the inputed String based on what is being search for. Time Complexity of O(n) where n is the number of Product objects in the Array List. Straightforward and simple traversal for this class.