First version for Shopping Cart Class
This is the shopping cart class that will keep track of the products in a users shopping cart. It uses an ArrayList<Product> to keep track of any added products and removed products from the shopping cart. There is also a separate method to calculate the total price of the products in the shopping cart. There is also another method to get what products are in the shopping cart. This class will need to be implemented into the User class since each user cannot share a shopping cart online.