Fixed the ShoppingCart
So I noticed that with my second constructor in the ShoppingCart Class a new Product Catalog object was being made without actually storing the contents of the current product catalog that was already made. So it kept printing empty, $0.0, and the product not found in catalog because of that constructor. So I changed up the constructor in the User class by making another one that will also take in the current product catalog. The second User constructor will be used to register the User solely for its username and password and not the shopping cart since that will be not needed in the UserManagement class.