First version of recommendation system class
This is the starting version of the RecommendationSystem class before testing in the Main class. The RecommendationSystem class uses dynamic programming, specifically, memoization to store any similar products and provide a Product AraryList of recommended products for the user based on product description. A recursive call is used to generate recommendations and go through the product catalog to find each product with a matching description. It is meant to avoid redundant results or input. It has a base case where the product description might be empty or null, if so, an empty array list is returned.