Fixing the Infinite Recursion
So I ended up with a stack overflow error since there was a cycle or infinite recursion happening with all the products in the product catalog as we searched for products by description. So I used a HashSet called visited to keep track of the products I had previously visited so an infinite recursion would not happen.