First Made version of AVL Tree
This is just a starting version for the AVL Tree class which will be used for the Order Class. We have methods to get the balance of the tree, doing rotations to the left or right, print statements, and an insert method that updates the height and rebalances the tree. 
Memory Usage: 366 M of 2048 M