Order Java Class Update 2
Updated the Order class to be an abstract class that will be instantiated in the AVLTree class. I had the Order class implement Comparable<Order> which is a built in Java interface class that does an ordering of a class or Object in it. I can use a compareTo method to find the order of the node for balancing the tree. 
The Order class also takes in a timestamp of the current time in milliseconds. 
Memory Usage: 326 M of 2048 M