3rd and hopefully Final Version of Order
Alright so I made some adjustments to the Order class based on what was done for the AVLTree and OrderManager class. So a toString method was made to print out the full Order (User name, Products in their Cart, Total Price of Products, and Order Id) so it would print out in an organized format. I didn't want the class t be abstract since I didn't need the Comparator Interface Class anymore. I took out the timestamp variable since it felt unnecessary when I was testing out the code. 
Memory Usage: 365 M of 2048 M