3rd Version of AVLTree Class
So a lot of changes were made in this class so let me explain. I changed the values I used for the key and node.data since I wanted to compare the prices of each order and set up the AVL Tree to print out the orders in ascending order. I used the getTotalPrice() method from the Order class to get the total price of each order and I changed the data into Object data instead of T data since I no longer had Object as an abstract class and I didn't need any comparing methods. A getRoout(0 method was made for the OrderManager class where the AVLTree will be used. I made a few statements that dealt with any null values so the program would not stop to tell me about a NullPointerException in the leftRotate or rightRotate classes. 
Memory Usage: 369 M of 2048 M