OrderManager
This class uses an object from the AVLTree class to addOrders, removeOrders, getOrderById, and searchOrderById from the Main class. I have statements to handle any null values, a print Order statement to print all the orders from the AVLTree with in-order Traversal. From the AVLTree root node I find the Order using the Order Id attribute which is unique to each Order. I anticipate the Order to also not be the root node, so I recursively search through the nodes in the left and right subtree if the Order Id is not similar to the Order Id of the root node. 
Memory Usage: 322 M of 2048 M