UserManagement.java first trial
This is the UserManagement Class that handles the different user accounts. I use two hashMaps for the user accounts and userCustomerIDMap to map the inputed username to its customerID (used for authenticating an account). 
We have a void registerUser method that takes the username and password and creates a new User with it. Dynamic programming is used to consider scenarios where there is an already existing username and creating the new user account and storing the user data into the two hashMaps.
If a username already exists, a printed statement will appear letting the user know that the username already exists and to choose a new username and then exiting out of the method. 

If the username does not exist it will store the user's username, hashedPassword, and customerID into the userMap hashMap. Then in the userNameToCustomerIDMap it will store the username, and customerID from the newUser object.

The authenticateUser boolean method will return a true or false statement if the user was authenticated or not. If the username is found with its corresponding customerID in the usernameToCustomerIDMap then it will get the customerID, password, User object from the HashMap and check if the correct password corresponds to the username with the hashedPassword stored in the userMap HashMap. 
If the username could not be found in the usernameToCustomerIDMap then the method will immediately return false.
Memory Usage: 365 M of 2048 M