Testing for Final Main Code User Register
I am currently testing the functionality of User register with a while loop that will continue until a successful User Account is made and registered into UserManagement. I am still in the progress of authenticating Users so this is just a benchmark for the registering User portion for making the final Main class.
Expected output:
There will be a prompt for the User to choose between authenticating, making a new User account, or exiting from the platform. If the User chooses 1, they will be prompted to input a username and password in a while loop. That loop will run while a success variable is false. If the username and password are new, the User is registered, success becomes true, and they exit the while loop. Or, if the username already exists the loop will tell them to make a different username and continue until a new User is successfully registered. 
Time Elapsed: 2 sec, 379 ms
Memory Usage: 436 M of 2048 M